#Showing all resources from backend on Resources page as a recyclerView

+++ Disaster-Response-Platform/mobile/DisasterResponsePlatform/app/src/main/java/com/example/disasterresponseplatform/data/database/resource/Resource.kt
@@ -12,7 +12,7 @@ import com.example.disasterresponseplatform.data.enums.ResourceCondition
 data class Resource(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = NeedResourceCols.id)
-    val ID: Int?, // if user does not enter ID, it generates it automatically
+    val ID: Long?, // if user does not enter ID, it generates it automatically
     @ColumnInfo(name = NeedResourceCols.creatorName)
     val creatorName: String,
     @ColumnInfo(name = NeedResourceCols.condition)
